Updated conf.py to allow greater flexibility in group_names and Django Mappings
In Okta there is a 100 AD group limit to the number of groups returned for a user.  Also often you will create a custom group for your application.  To limit the number of AD Groups, I created a custom name and populated it using a regex to return only names like "app_XYZ_*"  Thus getting round the 100 name limit.  But we need to pass in the group_name if not the default.

Django User Model Mapping (To allow users to use custom Okta fields to map to Django)