Update pre commit (#27)
Update pre-commit to latest version 3.1.1, and new versions of pre-commit tools. This also led to updates of dev dependencies to keep the environment aligned with pre-commit.

Updated PyJWT to support any minor version of major version 2.

Add a timeout to requests made by the requests library to Okta. Defaults to 10s but added a new setting to allow it to be configured.

Minor style improvements across everything.

Update test mocks to return proper certificates to testing properly validates JWTs.

---------
Authored-by: Matt Magin <matt@magin.id.au>